pyproject.toml
[project]
name = "prepforge"
version = "0.2.0"
description = "AI Exam Preparation System with TUI, GUI, and Streamlit"
authors = [
    { name = "Himanshu Arora" }
]
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"

license = "MIT"
license-files = ["LICENSE"]

dependencies = [
    "transformers",
    "peft",
    "rich",
    "prompt_toolkit"
]

classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
gui = ["PyQt6"]
web = ["streamlit"]
train = ["datasets", "trl"]
gpu = ["torch"]
fast = ["unsloth"]
gguf = ["llama-cpp-python"]

[project.scripts]
prepforge = "major_project.cli:main"

[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["major_project*"]
